+++ jax/experimental/export/__init__.py
@@ -23,11 +23,12 @@ from jax.experimental.export._export import (
     DisabledSafetyCheck,
     default_lowering_platform,
 
-    args_specs,  # TODO: move to shape_poly
+    args_specs,  # TODO: deprecate
 )
 from jax.experimental.export.shape_poly import (
     is_symbolic_dim,
     symbolic_shape,
+    symbolic_args_specs,
 )
 from jax.experimental.export.serialization import (
     serialize,