+++ jax/experimental/export/_export.py
@@ -345,7 +345,9 @@ def symbolic_shape(
   return shape_poly.symbolic_shape(shape_spec, like=like)
 
 def shape_and_dtype_jax_array(a) -> tuple[Sequence[int | None], DType]:
-  """Returns the shape and dtype of a jax.Array."""
+  """Returns the shape and dtype of a jax.Array or a j"""
+  if isinstance(a, jax.ShapeDtypeStruct):
+    return a.shape, a.dtype
   aval = core.raise_to_shaped(core.get_aval(a))
   return aval.shape, aval.dtype
 
@@ -354,55 +356,15 @@ def args_specs(
     polymorphic_shapes,  # prefix pytree of strings
     get_shape_and_dtype=shape_and_dtype_jax_array,
 ):
-  """Constructs a pytree of jax.ShapeDtypeSpec arguments specs for `export`.
-
-  Args:
-    args: a pytree of arguments
-    polymorphic_shapes: should be `None` (all arguments are monomorphic),
-      a single string (applies to all arguments), or a pytree matching a prefix
-      of the `args`.
-      See [how optional parameters are matched to
-      arguments](https://jax.readthedocs.io/en/latest/pytrees.html#applying-optional-parameters-to-pytrees).
-
-      Note that this function does not ensure that the provided `args` shapes
-      are compatible with `polymorphic_shapes`. The `.shape` of the `args` are
-      used only to fill-in placeholders from `polymorphic_shapes`.
-
-      See docstring of `symbolic_shape` and
-      [the README](https://github.com/google/jax/blob/main/jax/experimental/jax2tf/README.md#shape-polymorphic-conversion)
-      for more details.
-    get_shape_and_dtype: a function that given an argument extracts a tuple
-      of a shape and a dtype.
-
-  Returns: a pytree of jax.ShapeDTypeStruct matching `args`.
-  """
-  args_flat, args_tree = tree_util.tree_flatten(args)
-
-  shapes_and_dtypes = tuple(map(get_shape_and_dtype, args_flat))
-  shapes, dtypes = util.unzip2(shapes_and_dtypes)
-
-  if isinstance(args, tuple) and isinstance(polymorphic_shapes, list):
-    # TODO: Remove backward-compatibility workaround
-    polymorphic_shapes_ = tuple(polymorphic_shapes)
-  else:
-    polymorphic_shapes_ = polymorphic_shapes
-
-  try:
-    polymorphic_shapes_flat = tree_util.broadcast_prefix(
-        polymorphic_shapes_, args,
-        is_leaf=lambda x: x is None)
-  except ValueError:
-    e, *_ = tree_util.prefix_errors(
-        polymorphic_shapes_, args,
-        is_leaf=lambda x: x is None)
-    raise e("jax_export polymorphic_shapes") from None
-
-  # Now add in the polymorphic shapes
-  args_specs_flat = (
-      jax.ShapeDtypeStruct(symbolic_shape(spec, like=s), t)
-      for s, t, spec in zip(shapes, dtypes, polymorphic_shapes_flat))
-
-  return args_tree.unflatten(args_specs_flat)
+  # TODO: deprecated in January 2024, to be removed.
+  warnings.warn(
+      "export.args_specs is deprecated in favor of export.symbolic_args_specs",
+      DeprecationWarning, stacklevel=2)
+  if get_shape_and_dtype is not shape_and_dtype_jax_array:
+    # This was needed in some older jax2tf implementations
+    args = tree_util.tree_map(lambda a: jax.ShapeDtypeStruct(* get_shape_and_dtype(a)),
+                              args)
+  return shape_poly.symbolic_args_specs(args, polymorphic_shapes)
 
 
 def _keep_main_tokens(serialization_version: int) -> bool: